feat: implement B1 — Firebase Function proxy for MCP server
- functions/src/mcp/proxy.ts: mcpProxy onRequest function that fetches a
  Google identity token for Cloud Run IAM and forwards the user MAPLE token
  in X-Maple-Authorization; Cloud Run stays --no-allow-unauthenticated
- functions/src/index.ts: export mcpProxy
- firebase.json: add hosting rewrite /api/mcp → mcpProxy (us-central1)
- Remove pages/api/mcp.ts — static export doesn't support API routes
- google-auth-library added to functions/package.json
- AiTools.tsx: add TODO to update connection config once proxy is live
- roles/run.invoker granted to App Engine default SA (Firebase Functions)
