+++ functions/package.json
@@ -13,16 +13,17 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "@google-cloud/aiplatform": "^3.9.0",
     "@google-cloud/firestore": "^5.0.2",
     "@google-cloud/pubsub": "^3.0.1",
     "assemblyai": "^4.9.0",
     "axios": "^0.25.0",
     "date-fns": "^2.30.0",
     "firebase-admin": "^12.0.0",
-    "@google-cloud/aiplatform": "^3.9.0",
     "firebase-functions": "^5.1.1",
     "fluent-ffmpeg": "^2.1.3",
     "fuse.js": "6.5.3",
+    "google-auth-library": "^10.6.2",
     "handlebars": "^4.7.8",
     "js-sha256": "^0.11.0",
     "jsdom": "^26.0.0",
