+++ functions/src/index.ts
@@ -62,6 +62,8 @@ export { transcription } from "./webhooks"
 
 export * from "./triggerPubsubFunction"
 
+export { mcpProxy } from "./mcp/proxy"
+
 // Export the health check last so it is loaded last.
 export * from "./healthCheck"
